@@ -1,14 +1,14 @@
-import 'package:new_architecture/services/epoch_service.dart';
-import 'package:new_architecture/services/factory_service.dart';
-import 'package:new_architecture/services/information_service.dart';
-import 'package:new_architecture/ui/bottom_nav/bottom_nav_example.dart';
-import 'package:new_architecture/ui/bottom_nav/favorites/favorites_viewmodel.dart';
-import 'package:new_architecture/ui/bottom_nav/history/history_viewmodel.dart';
-import 'package:new_architecture/ui/details/details_view.dart';
-import 'package:new_architecture/ui/form/example_form_view.dart';
-import 'package:new_architecture/ui/home/home_view.dart';
-import 'package:new_architecture/ui/nonreactive/nonreactive_view.dart';
-import 'package:new_architecture/ui/stream_view/stream_counter_view.dart';
+import 'package:example/services/epoch_service.dart';
+import 'package:example/services/factory_service.dart';
+import 'package:example/services/information_service.dart';
+import 'package:example/ui/bottom_nav/bottom_nav_example.dart';
+import 'package:example/ui/bottom_nav/favorites/favorites_viewmodel.dart';
+import 'package:example/ui/bottom_nav/history/history_viewmodel.dart';
+import 'package:example/ui/details/details_view.dart';
+import 'package:example/ui/form/example_form_view.dart';
+import 'package:example/ui/home/home_view.dart';
+import 'package:example/ui/nonreactive/nonreactive_view.dart';
+import 'package:example/ui/stream_view/stream_counter_view.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked/stacked_annotations.dart';
 import 'package:stacked_services/stacked_services.dart';
