@@ -9,7 +9,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
-import 'package:stacked/stacked_annotations.dart';
 import 'package:stacked_services/stacked_services.dart';
 
 import '../ui/bottom_nav/bottom_nav_example.dart';
@@ -64,7 +63,7 @@ class StackedRouter extends RouterBase {
     },
     StreamCounterView: (data) {
       return MaterialPageRoute<dynamic>(
-        builder: (context) => StreamCounterView(),
+        builder: (context) => const StreamCounterView(),
         settings: data,
       );
     },
