@@ -1,17 +1,3 @@
 import 'package:stacked/stacked.dart';
 
-class BottomNavExampleViewModel extends BaseViewModel {
-  int _currentTabIndex = 0;
-  int get currentTabIndex => _currentTabIndex;
-
-  bool _reverse = false;
-  bool get reverse => _reverse;
-
-  void setTabIndex(int value) {
-    if (value < _currentTabIndex) {
-      _reverse = true;
-    }
-    _currentTabIndex = value;
-    notifyListeners();
-  }
-}
+class BottomNavExampleViewModel extends IndexTrackingViewModel {}
