@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:new_architecture/app/app.locator.dart';
+import 'package:example/app/app.locator.dart';
 import 'package:stacked/stacked.dart';
 
 import '../../../app/app.locator.dart';
@@ -18,7 +18,7 @@ class FavoritesView extends StatelessWidget {
           body: Center(
               child: Text(
             viewModel.counter.toString(),
-            style: TextStyle(fontSize: 30),
+            style: const TextStyle(fontSize: 30),
           ))),
       viewModelBuilder: () => exampleLocator<FavoritesViewModel>(),
       onModelReady: (viewModel) => viewModel.setCounterTo999(),
