@@ -1,10 +1,9 @@
 import 'package:stacked/stacked.dart';
 
 class FavoritesViewModel extends BaseViewModel {
-  
-
   int _counter = 0;
   int get counter => _counter;
+
   void incrementCounter() {
     _counter++;
     notifyListeners();
