@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:new_architecture/app/app.router.dart';
+import 'package:example/app/app.router.dart';
 import 'package:stacked_services/stacked_services.dart';
 
 class DetailsView extends StatelessWidget {
@@ -9,8 +9,8 @@ class DetailsView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    NavigationService _navigationService = NavigationService();
-    _navigationService.navigateTo(
+    NavigationService navigationService = NavigationService();
+    navigationService.navigateTo(
       Routes.detailsView,
       arguments: DetailsViewArguments(name: 'FilledStacks'),
     );
