@@ -1,26 +1,24 @@
 import 'package:flutter/material.dart';
-import 'package:new_architecture/datamodels/human.dart';
+import 'package:example/datamodels/human.dart';
 import 'package:stacked/stacked.dart';
 
 class FullNameWidget extends ViewModelWidget<Human> {
+  const FullNameWidget({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context, Human model) {
     return Row(
       children: <Widget>[
-        Container(
-          child: Text(
-            model.name!,
-            style: TextStyle(fontWeight: FontWeight.bold, fontSize: 30),
-          ),
+        Text(
+          model.name!,
+          style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 30),
         ),
-        SizedBox(
+        const SizedBox(
           width: 50,
         ),
-        Container(
-          child: Text(
-            model.surname!,
-            style: TextStyle(fontWeight: FontWeight.bold, fontSize: 30),
-          ),
+        Text(
+          model.surname!,
+          style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 30),
         ),
       ],
     );
