@@ -1,9 +1,10 @@
 class FormValidators {
   static String? passwordValidator(String? value) {
-    if (value == null || value.isEmpty)
+    if (value == null || value.isEmpty) {
       return "Password should not be empty";
-    else
+    } else {
       return null;
+    }
   }
 
   static String? emailValidator(String? value) {
