@@ -1,6 +1,6 @@
-import 'package:new_architecture/app/app.locator.dart';
-import 'package:new_architecture/app/app.logger.dart';
-import 'package:new_architecture/app/app.router.dart';
+import 'package:example/app/app.locator.dart';
+import 'package:example/app/app.logger.dart';
+import 'package:example/app/app.router.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
@@ -9,11 +9,8 @@ class HomeViewModel extends BaseViewModel {
   final NavigationService _navigationService =
       exampleLocator<NavigationService>();
 
-  HomeViewModel() {
-    log.d('created');
-  }
-
   String title = 'default';
+  int counter = 0;
 
   void navigate() {
     _navigationService.navigateTo(
@@ -31,11 +28,10 @@ class HomeViewModel extends BaseViewModel {
 
   void initialise() {
     log.i('initialise');
-    title = 'initialised';
+    title = 'Initialised';
     notifyListeners();
   }
 
-  int counter = 0;
   void updateTitle() {
     counter++;
     title = '$counter';
