@@ -1,5 +1,5 @@
-import 'package:new_architecture/app/app.locator.dart';
-import 'package:new_architecture/services/information_service.dart';
+import 'package:example/app/app.locator.dart';
+import 'package:example/services/information_service.dart';
 import 'package:stacked/stacked.dart';
 
 class WidgetTwoViewModel extends ReactiveViewModel {
