@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:new_architecture/ui/startup/startup_viewmodel.dart';
+import 'package:example/ui/startup/startup_viewmodel.dart';
 import 'package:stacked/stacked.dart';
 
 class StartupView extends StatelessWidget {
@@ -9,7 +9,7 @@ class StartupView extends StatelessWidget {
   Widget build(BuildContext context) {
     return ViewModelBuilder<StartupVieWModel>.reactive(
       viewModelBuilder: () => StartupVieWModel(),
-      builder: (context, viewModel, child) => Scaffold(
+      builder: (context, viewModel, child) => const Scaffold(
         body: Center(
           child: Text('Startup View'),
         ),
