@@ -1,18 +1,41 @@
 import 'package:flutter/material.dart';
-import 'package:new_architecture/ui/stream_view/stream_counter_viewmodel.dart';
+import 'package:example/ui/stream_view/stream_counter_viewmodel.dart';
 import 'package:stacked/stacked.dart';
 
 class StreamCounterView extends StatelessWidget {
+  const StreamCounterView({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return ViewModelBuilder<StreamCounterViewModel>.reactive(
       builder: (context, viewModel, child) => Scaffold(
-        body: Center(
-          child: Text(viewModel.title),
+        appBar: AppBar(
+          title: const Text('Stream Counter View'),
+          centerTitle: true,
         ),
-        floatingActionButton: MaterialButton(
-          child: Text('Change Stream Srouces'),
-          onPressed: viewModel.changeStreamSources,
+        body: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              ElevatedButton(
+                onPressed: viewModel.changeStreamSources,
+                child: const Text('Change Stream Sources'),
+              ),
+              const SizedBox(height: 10),
+              Text(
+                viewModel.title,
+                textAlign: TextAlign.center,
+              ),
+              const SizedBox(height: 20),
+              Text(
+                viewModel.streamSource,
+                textAlign: TextAlign.center,
+              ),
+            ],
+          ),
+        ),
+        floatingActionButton: FloatingActionButton(
+          onPressed: viewModel.navigateToNewView,
         ),
       ),
       viewModelBuilder: () => StreamCounterViewModel(),
