@@ -1,12 +1,18 @@
-import 'package:new_architecture/app/app.locator.dart';
-import 'package:new_architecture/services/epoch_service.dart';
+import 'package:example/app/app.locator.dart';
+import 'package:example/app/app.router.dart';
+import 'package:example/services/epoch_service.dart';
 import 'package:stacked/stacked.dart';
+import 'package:stacked_services/stacked_services.dart';
 
 class StreamCounterViewModel extends StreamViewModel<int> {
+  final _navigationService = exampleLocator<NavigationService>();
+
   String get title => 'This is the time since epoch in seconds \n $data';
 
   late Stream<int> _currentSource;
   bool isSlowEpochNumbers = true;
+
+  String get streamSource => isSlowEpochNumbers ? 'Slow' : 'Fast';
 
   StreamCounterViewModel() {
     _setSource();
@@ -37,5 +43,9 @@ class StreamCounterViewModel extends StreamViewModel<int> {
     isSlowEpochNumbers = !isSlowEpochNumbers;
     _setSource();
     notifySourceChanged();
+  }
+
+  void navigateToNewView() {
+    _navigationService.navigateTo(Routes.exampleFormView);
   }
 }
