@@ -1,8 +1,8 @@
-import 'package:stacked/src/code_generation/router/route_data.dart';
-import 'package:stacked/src/code_generation/router/route_def.dart';
+import 'route_data.dart';
+import 'route_def.dart';
 
 import 'package:flutter/widgets.dart';
-import 'package:stacked/src/code_generation/router/route_matcher.dart';
+import 'route_matcher.dart';
 
 typedef StackedRouteFactory = Route<dynamic> Function(RouteData data);
 typedef RouterBuilder<T extends RouterBase> = T Function();
@@ -15,12 +15,11 @@ abstract class RouterBase {
   Set<String> get allRoutes => routes.map((e) => e.template).toSet();
 
   Route<dynamic>? onGenerateRoute(RouteSettings settings, [String? basePath]) {
-    assert(routes != null);
-    assert(settings != null);
     var match = findMatch(settings);
     if (match != null) {
       if (basePath != null) {
-        match = match.copyWith(name: _joinPath(basePath, match.name)) as RouteMatch?;
+        match = match.copyWith(name: _joinPath(basePath, match.name))
+            as RouteMatch?;
       }
 
       RouteData data;
